refactor: Reorganizing project structure in create
