+++ packages/create-preact-app/lib/resources/sw.js
