+++ packages/create-preact-app/lib/setup.js
@@ -1,5 +1,5 @@
 const spawn = require('cross-spawn-promise');
-const { hasCommand, warn } = require('../util');
+const { hasCommand, warn } = require('./util');
 
 const stdio = 'ignore';
